% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_TU_sparsity_auto.R
\name{best_keepgr_srr}
\alias{best_keepgr_srr}
\title{Different sparsity for comps}
\usage{
best_keepgr_srr(X, Y, n, nx, ny, lambda_kcv, groupx, groupy, keepx_seq,
  keepy_seq, tol = 1e-10, max_iterations = 100)
}
\arguments{
\item{X}{Numeric matrix. Vectors will be coerced to matrix with \code{as.matrix} (if this is possible)}

\item{Y}{Numeric matrix. Vectors will be coerced to matrix with \code{as.matrix} (if this is possible)}

\item{n}{Integer. Number of joint PLS components. Must be positive!}

\item{nx}{Integer. Number of orthogonal components in \eqn{X}. Negative values are interpreted as 0}

\item{ny}{Integer. Number of orthogonal components in \eqn{Y}. Negative values are interpreted as 0}

\item{lambda_kcv}{Integer. Number of folds of CV}

\item{groupx}{Character. A vecter or character indicating group names of the variables. The order of group names must correspond to the order of the vairables in X.}

\item{groupy}{Character. A vecter or character indicating group names of the variables. The order of group names must correspond to the order of the vairables in Y.}

\item{keepx_seq}{Vector. A vector indicating how many groups to keep for CV in each of the joint component of X.}

\item{tol}{double. Threshold for power method iteration}

\item{max_iterations}{Integer, Maximum number of iterations for power method}

\item{keepy_gr}{Vector. A vector indicating how many groups to keep for CV in each of the joint component of Y.}
}
\description{
Different sparsity for comps
}
