% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsPLS.R
\name{best_lambda}
\alias{best_lambda}
\title{K-fold CV based on symmetrized prediction error to find lambda}
\usage{
best_lambda(X, Y, n, lambda_kcv, n_lambda = 6, tol = 1e-10,
  max_iterations = 100)
}
\arguments{
\item{X}{Numeric matrix. Vectors will be coerced to matrix with \code{as.matrix} (if this is possible)}

\item{Y}{Numeric matrix. Vectors will be coerced to matrix with \code{as.matrix} (if this is possible)}

\item{n}{Integer. Number of joint PLS components. Must be positive!}

\item{lambda_kcv}{Integer. k-fold cross-validation for lambda_x and lambda_y}

\item{n_lambda}{Integer. Number of lambda_x and lambda_y to be tested. Search will be done on a n*n grid}

\item{tol}{double. Threshold for power method iteration}

\item{max_iterations}{Integer, Maximum number of iterations for power method}
}
\value{
List with the best values of lambda_x and lambda_y, as well as a table containing the mean error of every combination calculated. 
\item{x}{Contains the best lambda_x value}
\item{y}{Contains the best lambda_y value}
\item{grid}{Contains all the mean error calculated for each combination}
}
\description{
Search for the combination of lambdas which minimise the symmetrical prediction error (\code{\link{loocv_combi}})
}
\details{
Note that this function can be easily parallelized (on Windows e.g. with the \code{parallel} package.).
If there are NAs in the CVerr component, this is due to an error in the fitting.
}
