% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crossval_O2PLS.R
\name{crossval_o2m}
\alias{crossval_o2m}
\title{Cross-validate procedure for O2PLS}
\usage{
crossval_o2m(X, Y, a = 1, a2 = 1, b2 = 1, kcv, nr_cores = 1,
  stripped = TRUE, p_thresh = 3000, q_thresh = p_thresh, tol = 1e-10,
  max_iterations = 100)
}
\arguments{
\item{X}{Numeric matrix. Vectors will be coerced to matrix with \code{as.matrix} (if this is possible)}

\item{Y}{Numeric matrix. Vectors will be coerced to matrix with \code{as.matrix} (if this is possible)}

\item{a}{Vector of positive integers. Denotes the numbers of joint components to consider.}

\item{a2}{Vector of non-negative integers. Denotes the numbers of X-specific components to consider.}

\item{b2}{Vector of non-negative integers. Denotes the numbers of Y-specific components to consider.}

\item{kcv}{Positive integer. Number of folds to consider. Note: \code{kcv=N} gives leave-one-out CV.}

\item{nr_cores}{Positive integer. Number of cores to use for CV. You might want to use \code{\link{detectCores}()}.}

\item{stripped}{Logical. Use the stripped version of o2m (usually when cross-validating)?}

\item{p_thresh}{Integer. If \code{X} has more than \code{p_thresh} columns, a power method optimization is used, see \code{\link{o2m2}}}

\item{q_thresh}{Integer. If \code{Y} has more than \code{q_thresh} columns, a power method optimization is used, see \code{\link{o2m2}}}

\item{tol}{double. Threshold for power method iteration}

\item{max_iterations}{Integer, Maximum number of iterations for power method}
}
\value{
List of class \code{"cvo2m"} with the original and sorted Prediction errors and the number of folds used.
}
\description{
Cross-validate procedure for O2PLS
}
\examples{
local({
X = scale(jitter(tcrossprod(rnorm(100),runif(10))))
Y = scale(jitter(tcrossprod(rnorm(100),runif(10))))
crossval_o2m(X, Y, a = 1:4, a2 = 1:2, b2 = 1:2, 
             kcv = 5, nr_cores = 1)
})

}

