% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crossval_OmicsPLS.R
\name{crossval_sparsity}
\alias{crossval_sparsity}
\title{Perform cross-validation to find the optimal number of variables/groups to keep for each joint component}
\usage{
crossval_sparsity(
  X,
  Y,
  n,
  nx,
  ny,
  nr_folds,
  keepx_seq = NULL,
  keepy_seq = NULL,
  groupx = NULL,
  groupy = NULL,
  tol = 1e-10,
  max_iterations = 100
)
}
\arguments{
\item{X}{Numeric matrix. Vectors will be coerced to matrix with \code{as.matrix} (if this is possible)}

\item{Y}{Numeric matrix. Vectors will be coerced to matrix with \code{as.matrix} (if this is possible)}

\item{n}{Integer. Number of joint PLS components. Must be positive.}

\item{nx}{Integer. Number of orthogonal components in \eqn{X}. Negative values are interpreted as 0}

\item{ny}{Integer. Number of orthogonal components in \eqn{Y}. Negative values are interpreted as 0}

\item{nr_folds}{Integer. Number of folds of CV}

\item{keepx_seq}{Numeric vector. A vector indicating how many variables/groups to keep for CV in each of the joint component of X. Sparsity of each joint component will be selected sequentially.}

\item{keepy_seq}{Numeric vector. A vector indicating how many variables/groups to keep for CV in each of the joint component of Y. Sparsity of each joint component will be selected sequentially.}

\item{groupx}{Vector. Used when \code{sparse = TRUE}. A vector of strings indicating group names of each X-variable. Its length must be equal to the number of variables in \eqn{X}. The order of group names must corresponds to the order of the variables.}

\item{groupy}{Vector. Used when \code{sparse = TRUE}. A vector of strings indicating group names of each Y-variable. The length must be equal to the number of variables in \eqn{Y}. The order of group names must corresponds to the order of the variables.}

\item{tol}{Double. Threshold for which the NIPALS method is deemed converged. Must be positive.}

\item{max_iterations}{Integer. Maximum number of iterations for the NIPALS method.}
}
\value{
A list with the following elements:
   \item{Best}{a vector giving for each join component the number of features
  to keep from `X` and `Y` that yield the highest covariance between the
  joint components of `X` and `Y` (elements `x1`, `y1`, `x2`, `y2`, etc),
  and the number of features to keep from `X` and `Y` yielding the highest
  covariance under the 1 standard error rule (elements `x_1sd1`, `y_1sd1`,
  `x_1sd2`, `y_1sd2`, etc).}
   \item{Covs}{a list, with as many elements as number of joint components (`n`).
  Each element is a matrix giving the average covariance between the joint
  components of `X` and `Y` obtained across the folds, for each tested values
  of `keepx` (columns) and of `keepy` (rows).}
   \item{SEcov}{a list, with as many elements as number of joint components (`n`).
  Each element is a matrix giving the standard error of the covariance
  between the joint components of `X` and `Y` obtained across the folds, for
  each tested values of `keepx` (columns) and of `keepy` (rows).}
}
\description{
Perform cross-validation to find the optimal number of variables/groups to keep for each joint component
}
