% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsPLS.R
\name{cv_lambda_checker}
\alias{cv_lambda_checker}
\title{Check if sparsity parameters satisfy input conditions in cross-validation functions}
\usage{
cv_lambda_checker(x, y, keepx_seq = NULL, keepy_seq = NULL)
}
\arguments{
\item{x}{Should be numeric matrix.}

\item{y}{Should be numeric matrix.}

\item{keepx_seq}{Input of \code{crossval_sparsity} function.}

\item{keepy_seq}{Input of \code{crossval_sparsity} function.}
}
\description{
Check if sparsity parameters satisfy input conditions in cross-validation functions
}
\details{
This function throws an error if sparsity parameters are not valid.
}
\keyword{internal}
