% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsPLS.R
\name{cv_lambda_checker_group}
\alias{cv_lambda_checker_group}
\title{Check if sparisity parameters satisfy input conditions in cv functions}
\usage{
cv_lambda_checker_group(groupx, groupy, keepx_seq = NULL, keepy_seq = NULL)
}
\arguments{
\item{groupx}{Vector. Input of \code{cv_sparsity} function.}

\item{groupy}{Vector. Input of \code{cv_sparsity} function.}

\item{keepx_seq}{Input of \code{cv_sparsity} function.}

\item{keepy_seq}{Input of \code{cv_sparsity} function.}
}
\description{
Check if sparisity parameters satisfy input conditions in cv functions
}
\details{
This function throws an error if sparsity parameters are not valid.
}
\keyword{internal}
