% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsPLS.R
\name{delta}
\alias{delta}
\title{Find the delta to make the L1 norm equal to the lambda}
\usage{
delta(x, lambda)
}
\arguments{
\item{lambda}{L1 norm constraint}
}
\value{
Delta value
\item{del}{If delta = 0 results in the L1 norm of x less than lambda, return 0; otherwise return the value which make the L1 norm of x equal to lambda}
}
\description{
Find the delta to make the L1 norm equal to the lambda
}
