% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crossval_OmicsPLS.R
\name{err_back}
\alias{err_back}
\title{Perform cross-validation to find the optimal number of groups to keep for each joint component}
\usage{
err_back(dat, index, p, q)
}
\arguments{
\item{dat}{matrix with numeric row/col names}

\item{index}{get from which(..., arr.ind = T)}
}
\description{
Perform cross-validation to find the optimal number of groups to keep for each joint component
}
