% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsPLS.R
\name{lambda_checker_group}
\alias{lambda_checker_group}
\title{Check if penalization parameters satisfy input conditions}
\usage{
lambda_checker_group(groupx, groupy, keepx, keepy)
}
\description{
Check if penalization parameters satisfy input conditions
}
\details{
This function throws an error if lambda is not within the range.
}
\keyword{internal}
