% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O2PLS.R
\name{loadings.o2m}
\alias{loadings.o2m}
\title{Extract the loadings from an O2PLS fit}
\usage{
loadings.o2m(object, loading_name = c("Xjoint", "Yjoint", "Xorth", "Yorth"),
  subset = 0, ...)
}
\arguments{
\item{object}{List. Should be of class \code{\link{o2m}}.}

\item{loading_name}{character string. One of \code{"X"} (for plotting the X loadings) or \code{"Y"} (for plotting the Y loadings).}

\item{subset}{subset of loading vectors to be extracted.}

\item{...}{For compatibility}
}
\value{
Loading matrix
}
\description{
This function extracts loading parameters from an O2PLS fit
}
\examples{
loadings(o2m(scale(-2:2),scale(-2:2*4),1,0,0))
}

