% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O2PLS.R
\name{loocv_combi}
\alias{loocv_combi}
\title{K-fold CV based on symmetrized prediction error}
\usage{
loocv_combi(X, Y, a = 1:2, a2 = 1, b2 = 1, fitted_model = NULL,
  func = o2m_stripped, app_err = F, kcv)
}
\arguments{
\item{X}{Numeric matrix.}

\item{Y}{Numeric matrix.}

\item{a}{Vector or integers. Contains the #joint components.}

\item{a2}{Vector or integers. Contains the number of orthogonal components in \eqn{X}.}

\item{b2}{Vector or integers. Contains the number of orthogonal components in \eqn{Y}.}

\item{fitted_model}{List. O2PLS model fit with \code{\link{o2m}}. Is used to calculate the apparent error without recalculating this fit.}

\item{func}{Function to fit the O2PLS model with. Only \code{\link{o2m}} and \code{\link{o2m_stripped}} are supported.}

\item{app_err}{Logical. Should the apparent error also be computed?}

\item{kcv}{Integer. The value of \eqn{k}, i.e. the number of folds. Choose \eqn{N} for LOO-CV.}
}
\value{
List with two numeric vectors:
\item{CVerr}{Contains the k-fold CV estimated RMSEP}
\item{Fiterr}{Contains the apparent error}
}
\description{
The prediction error of both \code{X~Xhat} and \code{Y~Yhat} are summed. This provides a symmetrized version of \code{\link{loocv}}.
}
\details{
Note that this function can be easily parallelized (on Windows e.g. with the \code{parallel} package.).
}
\examples{
loocv_combi(matrix(c(-2:2)),matrix(c(-2:2*4)),1,0,0,func=o2m,kcv=5)
}

