% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsPLS.R
\docType{package}
\name{OmicsPLS}
\alias{OmicsPLS}
\title{O2PLS: Two-Way Orthogonal Partial Least Squares}
\description{
This is based on work of (Trygg, Wold, 2003).
Includes the O2PLS fit, some misc functions and some cross-validation tools.
}
\section{Model and assumptions}{

\strong{Note that the rows of \code{X} and \code{Y} are the subjects and columns are variables.}
The number of columns may be different, but the subjects should be the same in both datasets.

The O2PLS model (Trygg & Wold, 2003) decomposes two datasets \eqn{X} and \eqn{Y} into three parts. 
\itemize{
 \item{1.} A joint part, representing the relationship between \eqn{X} and \eqn{Y}
 \item{2.} An orthogonal part, representing the unrelated latent variation in \eqn{X} and \eqn{Y} separately.
 \item{3.} A noise part capturing all residual variation.
}

See also the corresponding paper for interpretation (el Bouhaddani et al, 2016).
}

\section{Fitting}{

The O2PLS fit is done with \code{\link{o2m}}. 
For data \code{X} and \code{Y} you can run \code{o2m(X,Y,n,nx,ny)} for an O2PLS fit with \code{n} joint and \code{nx, ny} orthogonal components.
See the help page of \code{\link{o2m}} for more information on parameters.
There are four ways to obtain an O2PLS fit, depending on the dimensionality.
\itemize{
 \item{} For the not-too-high dimensional case, you may use \code{\link{o2m}} with default parameters. E.g. \code{o2m(X,Y,n,nx,ny)}.
 \item{} In case you don't want the fancy output, but only the parameters, you may add \code{stripped = TRUE} to obtain a stripped version of \code{o2m} which avoids calculating and storing some matrices. E.g. \code{o2m(X,Y,n,nx,ny,stripped=TRUE)}.
 \item{} For high dimensional cases defined by \code{ncol(X)>p_thresh} and \code{ncol(Y)>q_thresh} a Power-Method approach is used which avoids storing large matrices. E.g. \code{o2m(X,Y,n,nx,ny,p_thresh=3000,q_thresh=3000)}.
 The thresholds are by default both at 3000 variables.
 \item{} If you want a stripped version in the high dimensional case, add \code{stripped = TRUE}. E.g. \code{o2m(X,Y,n,nx,ny,stripped=TRUE,p_thresh=3000,q_thresh=3000)}.
}
}

\section{Obtaining results}{

After fitting an O2PLS model, by running e.g. \code{fit = o2m(X,Y,n,nx,ny)}, the results can be visualised.
Use \code{\link{plot}(fit,...)} to plot the desired loadings with/without ggplot2.
Use \code{\link{summary}(fit,...)} to see the relative explained variances in the joint/orthogonal parts.
Also plotting the joint scores \code{fit$Tt, fit$U} and orthogonal scores \code{fit$T_Yosc, fit$U_Xosc} are of help.
}

\section{Cross-validating}{
 
Determining the number of components \code{n,nx,ny} is an important task. For this we have two methods.
See \code{citation("OmicsPLS")} for our proposed approach for determining the number of components, implemented in \code{crossval_o2m_adjR2}!
\itemize{
 \item{} Cross-validation (CV) is done with \code{\link{crossval_o2m}} and \code{\link{crossval_o2m_adjR2}}, both have built in parallelization which relies on the \code{parallel} package.
 Usage is something like \code{crossval_o2m(X,Y,a,ax,ay)} where \code{a,ax,ay} are vectors of integers. See the help pages.
 \code{kcv} is the number of folds, with \code{kcv = nrow(X)} for Leave-One-Out CV. 
 \item{} For \code{crossval_o2m_adjR2} the same parameters are to be specified. This way of cross-validating is (potentially much)
 faster than the standard approach. 
}
}

\section{Misc}{

Also some handy tools are available
\itemize{
 \item{} \code{\link{orth}(X)} is a function to obtain an orthogonalized version of a matrix or vector \code{X}.
 \item{} \code{\link{ssq}(X)} is a function to calculate the sum of squares (or squared Frobenius norm) of \code{X}. See also \code{\link{vnorm}} for calculating the norm of each column in \code{X}.
 \item{} \code{\link{mse}(x, y)} returns the mean squared difference between two matrices/vectors. By default \code{y=0}.
}
}

\section{Citation}{

If you use the OmicsPLS R package in your research, please cite the corresponding paper:

\strong{Bouhaddani, S., Houwing-duistermaat, J., Jongbloed, G., Salo, P., Perola, M., & Uh, H.-W.} (2016).
\emph{Evaluation of O2PLS in Omics data integration.}
BMC Bioinformatics BMTL Supplement. doi:10.1186/s12859-015-0854-z

The bibtex entry can be obtained with command \code{citation("OmicsPLS")}.
Thank You!

The original paper proposing O2PLS is

\strong{Trygg, J., & Wold, S.} (2003). 
\emph{O2-PLS, a two-block (X-Y) latent variable regression (LVR) method with an integral OSC filter.} 
Journal of Chemometrics, 17(1), 53-64. http://doi.org/10.1002/cem.775
}

\author{
Said el Bouhaddani (\email{s.el_bouhaddani@lumc.nl}),
Jeanine Houwing-Duistermaat (\email{J.J.Houwing@lumc.nl}),
Geurt Jongbloed (\email{G.Jongbloed@tudelft.nl}),
Szymon Kielbasa (\email{S.M.Kielbasa@lumc.nl}),
Hae-Won Uh (\email{H.Uh@lumc.nl}).

Maintainer: Said el Bouhaddani (\email{s.el_bouhaddani@lumc.nl}).
}
\keyword{OmicsPLS}
