% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/O2PLS.R
\name{orth}
\alias{orth}
\title{Orthogonalize a matrix}
\usage{
orth(X)
}
\arguments{
\item{X}{Numeric vector or matrix.}
}
\value{
An orthogonalized representation of \eqn{X}
}
\description{
Orthogonalize a matrix
}
\details{
The definition of orthogonalizing a matrix is ambiguous. Here \code{orth(X)} relies on \code{link{svd}} to rotate \code{X} such that the first column is the unit norm first principal component etc.
If we write \eqn{X = UDV^T}, then \code{orth(X)} is equivalent to \eqn{UV^T}
An alternative, not implemented, way would be to orthonormalize the second column of \eqn{X} w.r.t. the first column, and the third column w.r.t. the first two columns etc.
The way we typically use it is to make a vector \eqn{X} unit norm.
}
\examples{
orth(c(3,4))
round(crossprod(orth(matrix(rnorm(500),100,5))),4)
}

