% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsPLS.R
\name{orth_vec}
\alias{orth_vec}
\title{Post-orthogonalization of a sparse loading vector with regard to a matrix}
\usage{
orth_vec(x, W)
}
\arguments{
\item{x}{sparse loading vector to be orthogonalized}

\item{W}{sparse loading matrix of the previous loading vectors}
}
\value{
A sparse loading vector
}
\description{
Post-orthogonalization of a sparse loading vector with regard to a matrix
}
