% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crossval_OmicsPLS.R
\name{plot_crossval_sparsity}
\alias{plot_crossval_sparsity}
\title{Plot sparsity cross-validation results}
\usage{
plot_crossval_sparsity(cv_res)
}
\arguments{
\item{cv_res}{List, result from \code{\link{crossval_sparsity}}.}
}
\value{
A ggplot.
}
\description{
The produced plot has one facet for each joint component. The x-axis
corresponds to the number of features retained from the X dataset to
construct the joint component, and the y-axis to the number of features
retained from the Y dataset to construct the joint component. The colour
of each point in the ith facet represents the average covariance obtained
between the joint ith components of the two datasets over the cross-validation
folds. The size of the points' shadow correspond to the covariance standard
error across the cross-validation folds. For each joint component, the
setting yielding the maximum average covariance is highlighted in orange,
and the one yielding the highest average covariance under the 1-SD rule in
red.
}
