% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crossval_OmicsPLS.R
\name{print.cvo2m}
\alias{print.cvo2m}
\title{Cross-validate procedure for O2PLS}
\usage{
\method{print}{cvo2m}(x, include_matrix = FALSE, ...)
}
\arguments{
\item{x}{List of class \code{"cvo2m"}, produced by \code{\link{crossval_o2m}}.}

\item{include_matrix}{Logical. Should the 3-d array with Prediction errors also be printed.}

\item{...}{For consistency.}
}
\description{
Cross-validate procedure for O2PLS
}
