% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O2PLS.R
\name{print.o2m}
\alias{print.o2m}
\title{Print function for O2PLS.}
\usage{
\method{print}{o2m}(x, ...)
}
\arguments{
\item{x}{An O2PLS fit (an object of class o2m)}

\item{...}{For consistency}
}
\description{
This function is the print method for an O2PLS fit
}

