% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsPLS.R
\name{print.summary.o2m}
\alias{print.summary.o2m}
\title{Prints the summary of an O2PLS fit}
\usage{
\method{print}{summary.o2m}(x, ...)
}
\arguments{
\item{...}{For compatibility}
}
\description{
Readable output is given in the form of percentages of variances explained.
}
\examples{
summary(o2m(scale(-2:2),scale(-2:2*4),1,0,0))
}
\keyword{internal}
