% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsPLS.R
\name{quadr}
\alias{quadr}
\title{Solve quadratic}
\usage{
quadr(x, lambda = lambda)
}
\arguments{
\item{lambda}{L1 norm constraint}
}
\value{
Solution to a quadratic function
\item{neg_root}{The smaller root of the two real positive solutions}
}
\description{
Solve quadratic
}
