% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/O2PLS.R
\name{rmsep}
\alias{rmsep}
\title{Root MSE of Prediction}
\usage{
rmsep(Xtst, Ytst, fit)
}
\arguments{
\item{Xtst}{Numeric vector or matrix.}

\item{Ytst}{Numeric vector or matrix.}

\item{fit}{\code{\link{o2m}} fit (on data without \code{Xtst} and \code{Ytst}).}
}
\value{
Mean squares difference between predicted Y and true Y
}
\description{
Calculates the Root MSE of prediction on test data. *Expected* to work in combination with \code{\link{loocv}}.
}
\details{
This function is the building block for \code{\link{loocv}}, as it produced the prediction error for test (left out) data.
}

