% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsPLS.R
\name{scores.o2m}
\alias{scores.o2m}
\title{Extract the scores from an O2PLS fit}
\usage{
scores.o2m(x, which_part = c("Xjoint", "Yjoint", "Xorth", "Yorth"),
  subset = 0, ...)
}
\arguments{
\item{x}{Object of class \code{o2m}}

\item{which_part}{character string. One of the following: 'Xjoint', 'Yjoint', 'Xorth' or 'Yorth'.}

\item{subset}{subset of scores vectors to be extracted.}

\item{...}{Ignored for now.}
}
\description{
This function extracts score matrices from an O2PLS fit
}
\seealso{
\code{\link{loadings}}
}
