% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsPLS.R
\name{screeplot_o2m}
\alias{screeplot_o2m}
\title{Screeplot for OmicsPLS}
\usage{
screeplot_o2m(X, Y, npcs = 20, ...)
}
\arguments{
\item{X}{Numeric matrix. Vectors will be coerced to matrix with \code{as.matrix} (if this is possible)}

\item{Y}{Numeric matrix. Vectors will be coerced to matrix with \code{as.matrix} (if this is possible)}

\item{npcs}{Positive integer. Number of singular values to plot}

\item{...}{Further arguments. Currently unused, but present for compatibility.}
}
\value{
Predicted Data
}
\description{
Plots singular values of X, X'Y and Y
}
\examples{
screeplot_o2m(scale(matrix(rnorm(100*30),100)), scale(matrix(rnorm(100*30),100)), npcs=20)
}
