% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/O2PLS.R
\name{ssq}
\alias{ssq}
\title{Calculate Sum of Squares}
\usage{
ssq(X)
}
\arguments{
\item{X}{Numeric vector or matrix.}
}
\value{
The sum of squared elements of \code{X}
}
\description{
Calculate Sum of Squares
}
\details{
This is the Frobenius norm of \eqn{X}.
}
\examples{
ssq(1:5)
ssq(rnorm(1e5))/1e5
}

