% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O2PLS.R
\name{summary.o2m}
\alias{summary.o2m}
\title{Summary of an O2PLS fit}
\usage{
\method{summary}{o2m}(object, perc = TRUE, ...)
}
\arguments{
\item{object}{List. Should be of class \code{\link{o2m}}.}

\item{perc}{Logical, show percentages?}

\item{...}{For compatibility}
}
\value{
List with R2 values.
}
\description{
Until now only variational summary given by the R2's is outputted
}
\examples{
summary(o2m(matrix(-2:2),matrix(-2:2*4),1,0,0))
}

