% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsPLS.R
\name{sumR2_combi}
\alias{sumR2_combi}
\title{Symmetrized sum of prediction R^2}
\usage{
sumR2_combi(Xtst, Ytst, fit)
}
\arguments{
\item{Xtst}{Numeric vector or matrix.}

\item{Ytst}{Numeric vector or matrix.}

\item{fit}{\code{\link{o2m}} fit (on data without \code{Xtst} and \code{Ytst}).}
}
\value{
Sum of the R^2 for \code{Xtst} and \code{Ytst}
}
\description{
Calculates the symmetrized sum of prediction R^2.
}
\details{
This function is the building block for \code{\link{best_lambda}}, as it produced the prediction error for test (left out) data.

The predicion error R^2 of both \code{Xtst} and \code{Ytst} are calculated and summed.
}
