% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsPLS.R
\name{thresh}
\alias{thresh}
\title{Soft threshholding a vector}
\usage{
thresh(z, delta)
}
\arguments{
\item{z}{Numerical vector}

\item{delta}{Soft-thresholding parameter}
}
\value{
Soft-thresholded vector
}
\description{
Soft threshholding a vector
}
