% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsPLS.R
\name{thresh_n}
\alias{thresh_n}
\title{Soft threshholding a vector with respect to a number of variables}
\usage{
thresh_n(x, keepx)
}
\arguments{
\item{x}{Numerical vector}

\item{keepx}{How many non-zero}
}
\value{
Soft-thresholded vector
}
\description{
Soft threshholding a vector with respect to a number of variables
}
