% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/O2PLS.R
\name{vnorm}
\alias{vnorm}
\title{Norm of a vector or columns of a matrix}
\usage{
vnorm(x)
}
\arguments{
\item{x}{Numeric vector or matrix.}
}
\value{
(columnwise) Euclidian norm of \eqn{x}
}
\description{
Norm of a vector or columns of a matrix
}

