% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsPLS.R
\name{lambda_checker}
\alias{lambda_checker}
\title{Check if penalization parameters satisfy input conditions}
\usage{
lambda_checker(lambda_x, lambda_y, x, y)
}
\arguments{
\item{lambda_x}{Should be between 1 and square root of the number of variables of X.}

\item{lambda_y}{Should be between 1 and square root of the number of variables of Y.}
}
\description{
Check if penalization parameters satisfy input conditions
}
\details{
This function throws an error if lambda is not within the range.
}
\keyword{internal}
