% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/O2PLS.R
\docType{package}
\name{O2PLS}
\alias{O2PLS}
\alias{O2PLS-package}
\title{O2PLS: Two-Way Orthogonal Partial Least Squares}
\description{
This is based on work of (Trygg, Wold, 2003).
Includes the O2PLS fit, some misc functions and some cross-validation tools.
}
\section{Functions}{

The O2PLS fit is done with \code{\link{o2m}}.
Cross-validation is done with \code{\link{loocv}} or \code{\link{adjR2}}, the last has built in parallelization (when you use Windows!) which relies on the \code{parallel} package.

List of all functions:\itemize{
\item{\code{\link{adjR2}}}
\item{\code{\link{loocv}}}
\item{\code{\link{loocv_combi}}}
\item{\code{\link{mse}}}
\item{\code{\link{orth}}}
\item{\code{\link{o2m}}}
\item{\code{\link{o2m_stripped}}}
\item{\code{\link{rmsep}}}
\item{\code{\link{rmsep_combi}}}
\item{\code{\link{ssq}}}
\item{\code{\link{summary_o2m}}}
\item{\code{\link{vnorm}}}
}
}
\author{
Said el Bouhaddani (\email{s.el_bouhaddani@lumc.nl}),
Jeanine Houwing-Duistermaat (\email{J.J.Houwing@lumc.nl}),
Geurt Jongbloed (\email{G.Jongbloed@tudelft.nl}),
Hae-Won Uh (\email{H.Uh@lumc.nl}).

Maintainer: Said el Bouhaddani (\email{s.el_bouhaddani@lumc.nl}).
}
\keyword{O2PLS}

