% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/O2PLS.R
\name{orth}
\alias{orth}
\title{Orthogonalize a matrix}
\usage{
orth(X)
}
\arguments{
\item{X}{Numeric vector or matrix.}
}
\value{
An orthogonalized representation of \code{X}
}
\description{
Orthogonalize a matrix
}
\details{
The definition of orthogonalizing a matrix is not ambiguous. Here \code{X} is represented by all eigenvectors. The eigenvalues are set to 1.
An alternative, not implemented, way is to orthonormalize the second columns of \code{X} w.r.t. the first column, and the third column w.r.t. the first two columns etc.
}
\examples{
orth(c(3,4))
round(crossprod(orth(matrix(rnorm(500),100,5))),4)
}

