% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsPLS_o2m.R
\name{pow_o2m}
\alias{pow_o2m}
\title{NIPALS method for PLS2}
\usage{
pow_o2m(X, Y, n, tol = 1e-10, max_iterations = 1000)
}
\arguments{
\item{X}{Numeric matrix. Vectors will be coerced to matrix with \code{as.matrix} (if this is possible)}

\item{Y}{Numeric matrix. Vectors will be coerced to matrix with \code{as.matrix} (if this is possible)}

\item{n}{Integer. Number of joint PLS components. Must be positive.}

\item{tol}{Double. Threshold for which the NIPALS method is deemed converged. Must be positive.}

\item{max_iterations}{Integer. Maximum number of iterations for the NIPALS method.}
}
\description{
Performs power method for PLS2 loadings.
}
\seealso{
\code{\link{o2m}}
}
\keyword{internal}
