% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsPLS.R
\name{predict.o2m}
\alias{predict.o2m}
\title{Predicts X or Y}
\usage{
\method{predict}{o2m}(object, newdata, XorY = c("X", "Y"), ...)
}
\arguments{
\item{object}{List. Should be of class \code{\link{o2m}}.}

\item{newdata}{New data, which one of X or Y is specified in \code{XorY}.}

\item{XorY}{Character specifying whether \code{newdata} is X or Y.}

\item{...}{For compatibility}
}
\value{
Predicted Data
}
\description{
Predicts X or Y based on new data on Y or X
}
\details{
Prediction is done after correcting for orthogonal parts.
}
\examples{
predict(o2m(scale(1:10), scale(1:10), 1, 0, 0), newdata = scale(1:5), XorY = "X")
}
