% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsPLS.R
\name{rmsep}
\alias{rmsep}
\title{Root MSE of Prediction}
\usage{
rmsep(Xtst, Ytst, fit, combi = FALSE)
}
\arguments{
\item{Xtst}{Numeric vector or matrix.}

\item{Ytst}{Numeric vector or matrix.}

\item{fit}{\code{\link{o2m}} fit (on data without \code{Xtst} and \code{Ytst}).}

\item{combi}{Logical. Should the symmetrized MSE be used, i.e. add both MSEs. Not yet implemented, but see \code{\link{rmsep_combi}}}
}
\value{
Mean squares difference between predicted Y and true Y
}
\description{
Calculates the Root MSE of prediction on test data. Only tested to work inside \code{\link{loocv}}.
}
\details{
This function is the building block for \code{\link{loocv}}, as it produced the prediction error for test (left out) data.
}
