% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsPLS.R
\name{ssq}
\alias{ssq}
\title{Calculate Sum of Squares}
\usage{
ssq(X)
}
\arguments{
\item{X}{Numeric vector or matrix.}
}
\value{
The sum of squared elements of \eqn{X}
}
\description{
Calculate Sum of Squares
}
\details{
This is the Frobenius norm of \eqn{X}.
}
\examples{
ssq(tcrossprod(1:5))
ssq(rnorm(1e5))/1e5
}
