% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsPLS.R
\name{summary.o2m}
\alias{summary.o2m}
\title{Summary of an O2PLS fit}
\usage{
\method{summary}{o2m}(object, digits = 3, ...)
}
\arguments{
\item{object}{List. Should be of class \code{\link{o2m}}.}

\item{digits}{Integer, number of digits.}

\item{...}{For compatibility}
}
\value{
List with R2 values.
}
\description{
Until now only variational summary given by the R2's is outputted
}
\examples{
summary(o2m(scale(-2:2),scale(-2:2*4),1,0,0))

}
\seealso{
\code{\link[OmicsPLS:plot.o2m]{plot}}
}
