% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsPLS.R
\name{vnorm}
\alias{vnorm}
\title{Norm of a vector or columns of a matrix}
\usage{
vnorm(x)
}
\arguments{
\item{x}{Numeric vector or matrix.}
}
\value{
(columnwise) Euclidian norm of \eqn{x}
}
\description{
Norm of a vector or columns of a matrix
}
\examples{
vnorm(orth(1:5))
vnorm(matrix(1:9,3,3))^2 - colSums(matrix(1:9,3)^2)
}
